* commit '3fbc3fd7bd4f37c4fbbfdb61644e4fe3091f0509': (34 commits)
  Rename columns densityWidth, densityHeight, cropX, cropY, cropWidth, cropHeight to avoid using a physical naming strategy
  IMGSERVER-21 Fix count @Query for Oracle and ms-sql
  IMGSERVER-21 Fix count @Query for Oracle and ms-sql
  IMGSERVER-21 Lombok
  Fix file location
  IMGSERVER-21 Image 404 does not leak input stream anymore
  IMGSERVER-21 Clean dependencies some more
  IMGSERVER-21 fix creating a new image resolution
  IMGSERVER-21 fix image-resolutions in admin web
  IMGSERVER-21 Removed common-spring and moved to IdLookup from hibernate module
  IMGSERVER-21 Goodbye MockedLoader
  IMGSERVER-21 use pom.version
  IMGSERVER-21 Avoid using the platform-bom version here
  IMGSERVER-21 Lombok mania
  IMGSERVER-21 Remove old repositories
  IMGSERVER-21 ImageModification.id
  IMGSERVER-21 Migrate to @ConfigurationProperties
  IMGSERVER-21 Added queries for get methods on ImageModificationRepository
  IMGSERVER-21 ImageModification extends Persistable - moved ImageModificationRepository to JpaRepository
  IMGSERVER-21 Migrate to @ConfigurationProperties
  ...

+++ imageserver-core/src/test/java/com/foreach/imageserver/core/services/CropGeneratorTest.java
@@ -57,11 +57,11 @@ public class CropGeneratorTest
 		imageResolutionService.setImageResolutionManager( imageResolutionManager );
 
 		image = new Image();
-		image.setId( 1 );
+		image.setId( 1L );
 		image.setDimensions( new Dimensions( 4000, 2000 ) );
 
 		context = new ImageContext();
-		context.setId( 1 );
+		context.setId( 1L );
 
 		modification = null;
 		resolutionIdCounter = 1;
@@ -212,7 +212,7 @@ public class CropGeneratorTest
 
 	private void requestModification( int width, int height ) {
 		ImageResolution resolution = new ImageResolution();
-		resolution.setId( 1 );
+		resolution.setId( 1L );
 		resolution.setWidth( width );
 		resolution.setHeight( height );
 
@@ -222,7 +222,7 @@ public class CropGeneratorTest
 	private ImageModification createModification( int width, int height, int x, int y, int cropWidth, int cropHeight ) {
 		ImageModification modification = new ImageModification();
 		ImageResolution resolution = new ImageResolution();
-		resolution.setId( resolutionIdCounter++ );
+		resolution.setId( (long) resolutionIdCounter++ );
 		resolution.setWidth( width );
 		resolution.setHeight( height );
