+++ imageserver-core/src/test/java/com/foreach/imageserver/core/services/support/AbstractImageModficationSelectorTest.java
@@ -41,7 +41,7 @@ abstract class AbstractImageModficationSelectorTest
 		              selector.selectImageModification( image, candidates, requested ) );
 	}
 
-	protected ImageResolutionDto resolution( int resolutionId, int width, int height ) {
+	protected ImageResolutionDto resolution( Long resolutionId, int width, int height ) {
 		ImageResolutionDto r = new ImageResolutionDto();
 		r.setWidth( width );
 		r.setHeight( height );
